Make Snake::UpdateBody virtual
This is because ObstacleSnakes may not need to chekc if they have
collided with themself or if they do, they could have a custom
implication for the collision that is different from the other snake.

Also other snakes may have custom implication for self collision like
the AISnake which may not die but maybe shrink to a single cell as its
penalty for self collision.