Fix: Store obstacles, playerSnake and food as their actual derived forms
The error shows up when I tried to create copies of the obstacles and
playerSnake for prediction/simulation. Hence, I switched to storing
the constant references to the vector of ObstacleSnake, const ref to
PlayerSnake, and Food. I did the same for food even though for now I do
not need to make a copy of the food only using it in read-only form.