Comment-out AISnake::ReviewPath()
Paused development of this feature because of time limitation and the
complication involved.

For now, If the PlayerSnake changes direction, FindPath() will be called
to create a new path to the goal regardles of if the new direction of
the playerSnake will result in the AISnake head running into the playerSnake's
body.