Move the forward declaration of Game from snakes.h to snake.h
This is because I have now also made Game a friend of Snake class hence
I needed to add a forward declaration for it in snake.h. And since I
have included snake.h in snakes.h, have the forward declaration in snake.h
will be included in snakes.h.