BugFix: AISnake not seeking food and game still freezing and fan ramp-up
Added an attribute in the Game class to keep track of the previous
food position for use in detecting whan a snake has eaten the food so
as to conditionally call AISnake::FindPath() in AISnake::SetDirectiion()
or when playerSnake direction has changed and there is a need to update
the path of the aiSnake due to a possible collision in the current path
the aiSnake is following. Hence the FindPath() is called conditionally
only when food position and/or playerSnake direction has changed.

This fix reduced the freezing and fan rampup issues. I think I still saw
a freeze when the aiSnake got very long - not sure.