Remove the code to prevent OppositeDirection movement
This is because after some analysis, I think the OppositeDirection() is
too simplistic to handle preventing the snake reversing onto itself.
I remove it to leave it upto the AddNode() to prevent self-collision.
This is suppposed to be fine since the only reason I added the reversal
prevention code was to reduce the computation or exploring a direction
I could simple detemine to result in self-collision.