BugFix: obstacles some times collision with aiSnake when near food
Noticed that sometimes the aiSnake rest of body collides with the obstacles
when it is about to eat the food. It seams the pathfinding is not taking
into account the rest of the body cells collision with the obstacles at
timesSteps after eating the food. This makes it inavitable that the aiSnake
will sometimes collide with an obstacle snake when its size is greater than
one after eating food because even if the pathfinding after eating food tries
to avoid the collision, it might not be able to due to speed limitations.