Update CheckCollision() to remove tails of some ai_snake_path_planning_issue
Remove the tails of snakes that got to the food cell after the first snake. This is because of
how the tails of any snake that is in the food cell after an update are not removed in case there
are multiple snakes that also got to the food cell after that same update. This is to wait till
a tie break is applied here in Game::CheckCollision().