Add code to block obstacleCollision after aiSnake eats food
This is an overly cautious attempt at making sure that there is no collision of
obstacle snakes into the aiSnake's cells from the goal to tail after the aiSnake eats the food. Because of
this overly cautious approach, some more optimal paths may be discarded and it will also take more time to
find a path to goal, increasing latency of the pathfinding. I may still need to
investigate this attempt because I thought I saw the issue again after this
implementation.