BugFix: maxTimeStep in _predictedObstacleBlockedCells did not match tailToPastGoalTime
The way the two variables are calculated is slightly different due to initialy only
using ceil with tailToPastGoalTime. This might be the cause of the (if I am not mistaking)
obstacle collision after reaching goal that I was sometimes still seeing even after
implmenting an algorithm to prevent that.

I have also added the equallity sign in the for loop of AISnake::PredictObstacleBlockedCells()
to include the boundary maxTimeStep, matching how I stop checking when the tail
reach the cell just after the goal(need to properly reanalys this).