Call PredictSnakeBlockedCells() in FindPath() before search start
This also involves modifying PredictObstacleBlockedCells() to remove
the buffer timeStep calculation and addition to the maxTimeStep in
the function since PredictSnakeBlockedCells() take the pre-calculated
maxTimeStep for obstacle snakes blocked cell prediction. This also means
the calculation of the buffer is centralized to the begining of the search
process and only done once in the search process unlike before.